add Component class
