+++ src/main.js
@@ -11,6 +11,8 @@ filtersContainer.innerHTML = FILTERS_DATA.map(createFilter).join(``);
 
 const tasks = getTasks(INITIAL_CARDS_LENGTH);
 
+let isOpened = false;
+
 tasks.forEach((task) => {
   const taskComponent = new Task(task);
   const editTaskComponent = new TaskEdit(task);
@@ -18,21 +20,26 @@ tasks.forEach((task) => {
   tasksContainer.appendChild(taskComponent.render());
 
   taskComponent.onEdit = () => {
-    editTaskComponent.render();
-    tasksContainer.replaceChild(editTaskComponent.element, taskComponent.element);
-    taskComponent.unrender();
+    if (!isOpened) {
+      editTaskComponent.render();
+      tasksContainer.replaceChild(editTaskComponent.element, taskComponent.element);
+      taskComponent.unrender();
+      isOpened = true;
+    }
   };
 
   editTaskComponent.onSubmit = () => {
     taskComponent.render();
     tasksContainer.replaceChild(taskComponent.element, editTaskComponent.element);
     editTaskComponent.unrender();
+    isOpened = false;
   };
 
   editTaskComponent.onClose = () => {
     taskComponent.render();
     tasksContainer.replaceChild(taskComponent.element, editTaskComponent.element);
     editTaskComponent.unrender();
+    isOpened = false;
   };
 });
