+++ src/task-edit.js
@@ -1,8 +1,10 @@
 import utils from "./utils.js";
 import {MONTHS_NAMES, KeyCodes} from './constants.js';
+import Component from "./component";
 
-export default class TaskEdit {
+export default class TaskEdit extends Component {
   constructor(data) {
+    super();
     this._color = data.color;
     this._title = data.title;
     this._dueDate = data.dueDate;
@@ -11,7 +13,6 @@ export default class TaskEdit {
     this._repeatingDays = data.repeatingDays;
     this._isFavorite = data.isFavorite;
 
-    this._element = null;
     this._onSubmit = null;
     this._onClose = null;
 
@@ -230,10 +231,6 @@ export default class TaskEdit {
           </article>`;
   }
 
-  get element() {
-    return this._element;
-  }
-
   set onSubmit(fn) {
     this._onSubmit = fn;
   }
@@ -263,7 +260,7 @@ export default class TaskEdit {
     document.body.addEventListener(`keydown`, this._onCloseCase);
   }
 
-  _removeListeners() {
+  removeListeners() {
     this._element.querySelector(`.card__form`)
       .removeEventListener(`submit`, this._onSubmitCase);
     document.body.removeEventListener(`click`, this._onCloseCase);
@@ -274,10 +271,5 @@ export default class TaskEdit {
     this._element = utils.createElement(this.template);
     setTimeout(this.addListeners.bind(this), 0);
     return this._element;
-  }
-
-  unrender() {
-    this._removeListeners();
-    this._element = null;
   }
 }
